Changed all Department reference to Calendar. Renamed ‘Filter’ to ‘filter’
Also had an issue in ‘event-data.service’ where the dataStoreService calendars$ observable
was being used as the parameter for calendarId in the getEventsOnCalendar function. Since I need an object to
create the checkboxes, it was passing an object as the calendarId and causing the call to fail.

I assigned calendarId = calendarId[‘label’] as a temporary fix.