WIP: Implementing month view.
Moved the links for the views into the month view and event list components.

Added initialized output to handle highlighting the selected date in the month view when switching back from day view.

Refactored code to take a more Rx-like approach.

Added subjects to trigger completion of observables so the application can unsubscribe automatically when the view is destroyed.

Fixed a typo in a method name in the date service.